# Provide an int64 version of Quantity that is much faster