Resolves #25019 by adding an int64 version of Quantity that is aware of overflow and can spill to the inf.Dec version.  Heavily optimized all the paths for allocation and efficiency.

```
BenchmarkQuantityString-8        	10000000	       175 ns/op	      23 B/op	       2 allocs/op
BenchmarkQuantityStringBinarySI-8	10000000	       214 ns/op	      32 B/op	       2 allocs/op
BenchmarkQuantityMarshalJSON-8   	10000000	       156 ns/op	      35 B/op	       1 allocs/op
BenchmarkQuantityUnmarshalJSON-8 	10000000	       194 ns/op	       2 B/op	       1 allocs/op
BenchmarkParseQuantity-8         	10000000	       142 ns/op	       0 B/op	       0 allocs/op
BenchmarkCanonicalize-8          	20000000	        89.2 ns/op	       3 B/op	       0 allocs/op
BenchmarkQuantityCopy-8          	20000000	        55.0 ns/op	      48 B/op	       1 allocs/op
BenchmarkQuantityAdd-8           	50000000	        33.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkQuantityRound-8         	50000000	        25.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkScaledValueSmall-8      	30000000	        35.9 ns/op	       0 B/op	       0 allocs/op
BenchmarkScaledValueLarge-8      	 3000000	       511 ns/op	      48 B/op	       1 allocs/op
```

Known issues

* [ ] Does not respect the current 12E max value, since that can be represented as 12*10^18 without loss of precision
* [ ] Shifting the scale down to merge the denominator with numerator in Parse (which avoids going to inf.Dec), i.e. "1.2" becomes "12e-1".  Questionable.
* [ ] Implement protobuf as a string.

<!-- Reviewable:start -->
---
This change is [<img src="http://reviewable.k8s.io/review_button.svg" height="35" align="absmiddle" alt="Reviewable"/>](http://reviewable.k8s.io/reviews/kubernetes/kubernetes/25243)
<!-- Reviewable:end -->