@smarterclayton 2016-05-10
Impact to general conversion and encoding

```
$ benchcmp /tmp/old1 /tmp/new1
benchmark                                      old ns/op     new ns/op     delta
BenchmarkDecodeCodec-8                         74170         57099         -23.02%
BenchmarkDecodeCodecToInternalProtobuf-8       14317         10915         -23.76%
BenchmarkDecodeIntoExternalCodec-8             63496         51679         -18.61%
BenchmarkDecodeIntoInternalCodec-8             71734         57493         -19.85%
BenchmarkDecodeIntoJSON-8                      57149         51707         -9.52%
BenchmarkDecodeIntoJSONCodecGen-8              34977         28385         -18.85%
BenchmarkDecodeIntoProtobuf-8                  5680          4814          -15.25%
BenchmarkEncodeCodec-8                         21975         19642         -10.62%
BenchmarkEncodeCodecFromInternal-8             28122         23517         -16.38%
BenchmarkEncodeCodecFromInternalProtobuf-8     11683         10165         -12.99%
BenchmarkEncodeCodecProtobuf-8                 6601          5533          -16.18%
BenchmarkEncodeJSONMarshal-8                   23308         19384         -16.84%
BenchmarkEncodeProtobufGeneratedMarshal-8      5460          4606          -15.64%
BenchmarkNodeConversion-8                      9641          6050          -37.25%
BenchmarkNodeCopy-8                            4746          6109          +28.72%
BenchmarkPodConversion-8                       6875          5819          -15.36%
BenchmarkPodCopy-8                             4433          4627          +4.38%
BenchmarkReplicationControllerConversion-8     7181          7133          -0.67%
BenchmarkReplicationControllerCopy-8           4116          4194          +1.90%

benchmark                                      old allocs     new allocs     delta
BenchmarkDecodeCodec-8                         337            226            -32.94%
BenchmarkDecodeCodecToInternalProtobuf-8       136            82             -39.71%
BenchmarkDecodeIntoExternalCodec-8             276            204            -26.09%
BenchmarkDecodeIntoInternalCodec-8             338            227            -32.84%
BenchmarkDecodeIntoJSON-8                      168            96             -42.86%
BenchmarkDecodeIntoJSONCodecGen-8              257            185            -28.02%
BenchmarkDecodeIntoProtobuf-8                  68             53             -22.06%
BenchmarkEncodeCodec-8                         66             39             -40.91%
BenchmarkEncodeCodecFromInternal-8             97             61             -37.11%
BenchmarkEncodeCodecFromInternalProtobuf-8     79             46             -41.77%
BenchmarkEncodeCodecProtobuf-8                 47             20             -57.45%
BenchmarkEncodeJSONMarshal-8                   62             34             -45.16%
BenchmarkEncodeProtobufGeneratedMarshal-8      43             16             -62.79%
BenchmarkNodeConversion-8                      85             29             -65.88%
BenchmarkNodeCopy-8                            28             32             +14.29%
BenchmarkPodConversion-8                       54             36             -33.33%
BenchmarkPodCopy-8                             25             25             +0.00%
BenchmarkReplicationControllerConversion-8     38             38             +0.00%
BenchmarkReplicationControllerCopy-8           22             22             +0.00%
```