@smarterclayton 2016-05-11
Note that I have no comments anywhere in the code (in case there was a desire to restructure), which I'll work on once someone takes a first pass and offers opinions.