@wojtek-t 2016-05-12
OK - since @lavalamp wants to look into it anyway and I don't have strong opinions on it, I'm reassigning it to him for now.